Refactored into new alarm module.
